esp.c: remove CMD_TI deferred interrupt logic in esp_transfer_data()
The deferred interrupt logic in esp_transfer_data() is for when a SCSI command
has a DATA phase and ensures that the data is present in the buffer when the
command completion interrupt is raised.

In the case of a TI command there is no need to handle the command completion
interrupt within esp_transfer_data() since we always call into esp_do_dma() and
esp_do_nodma() where the behaviour depends upon the current SCSI phase.

Remove the deferred interrupt logic for the TI command in esp_transfer_data()
since it is already handled elsewhere.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>